refactor(code-snippet): changes code highlight folder location and mdx plugins
